move : DiscountAmount를 모델 패키지로 이동
